Proto*: fix Error macro conflict
utils/Logger.h contains a preprocessor macro Error() for easy logging.
However, the upstream protobuf/io/coded_stream.h also defines a private
function Error() in one of its classes that conflicts with this macro.

Thus, unfortunate include orders result in a build error. This commit
reorders the includes to prevent the error. It also defines an extra
`#undef Error` before including upstream protobuf to make the problem
more visible (a compiler error would not point to an obscure place
somewhere in protobuf but to the usage of the Error() macro within
Hyperion itself).

Fixes: #882