webserver/Cmake: fix out-of-tree build
When using a cmake directory that is not a subfolder of the repository
root folder the current path replacement logic for webresources does not
work. This commit makes the replacement pattern also relative to the
build directory.

Reproduce with:
git clone hyperion-repo
mkdir build; cd build
cmake ../hyperion-repo

Without this chance the build itself succeeds but browsing the
webinterface results in:
404 - Requested file: index.html

Related: #834