webconfig: better error description, if Kodi is not found
Even if the normal Kodi webinterface is activated and capable of
receiving JSON RPC calls (on port 8080), the js in Hyperion sets the
port to 9090 (normally only accessible via localhost). Since the js code
is executed on the users device, Kodi denies access even if Hyperion and
Kodi are installed on the same device.

This change helps in this situation by pointing to the correct Kodi setting.